# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/user/biometric/credentials": {
            "get": {
                "description": "返回当前用户已注册的所有生物特征凭证",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[管理] 获取生物特征凭证列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.BiometricCredentialListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/user/biometric/credentials/{id}": {
            "delete": {
                "description": "删除指定的生物特征凭证",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[管理] 删除生物特征凭证",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "凭证 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "凭证不存在",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user/current": {
            "get": {
                "description": "根据访问令牌获取当前登录用户信息（用户名、邮箱）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[管理] 获取当前用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功，返回用户信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UserInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user/password": {
            "put": {
                "description": "验证旧密码并设置新密码，修改成功后撤销所有现有令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[管理] 修改密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "旧密码错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user/profile": {
            "put": {
                "description": "修改用户名和邮箱",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[管理] 更新个人资料",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "更新资料请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "string"
                },
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "overhead": {
                    "type": "integer"
                }
            }
        },
...
        "user.BiometricCredentialItem": {
            "type": "object",
            "properties": {
                "aaguid": {
                    "description": "认证器型号",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间戳",
                    "type": "integer"
                },
                "device_name": {
                    "description": "设备名称",
                    "type": "string"
                },
                "id": {
                    "description": "凭证 ID",
                    "type": "string"
                },
                "last_used_at": {
                    "description": "最后使用时间戳（nil 表示未使用）",
                    "type": "integer"
                }
            }
        },
        "user.BiometricCredentialListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "凭证列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.BiometricCredentialItem"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "user.UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码",
                    "type": "string",
                    "minLength": 8
                },
                "old_password": {
                    "description": "旧密码",
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "user.UpdateProfileRequest": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                }
            }
        },
        "user.UserInfoResponse": {
            "type": "object",
            "properties": {
                "biometric_credential_count": {
                    "description": "生物特征凭证数量",
                    "type": "integer"
                },
                "biometric_enabled": {
                    "description": "是否启用生物特征",
                    "type": "boolean"
                },
                "email": {
                    "description": "邮箱地址",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
    }
}